fix: §6.2/TODO-5 线程安全全面加固 + TODO-2/TODO-3/TODO-1审查→无需修改
§6.2 跨线程原子性:
- fade_cnt/ramp_cnt/mute_hold: volatile int→LONG + InterlockedExchange/Decrement
- freeze_lms: volatile int→long + InterlockedExchange (fxnlms_mimo.h)
- callback_count: volatile int→LONG + InterlockedIncrement
- hw.active_count: int→volatile int (howling_detect.h)
- wc_cur: CrossFader期间跳过scene_ctrl_process, 防止回调读时被主线程覆盖
- wc_old: InterlockedExchange(fade_cnt)提供释放屏障, 已安全
- fx.wc主线程读/监控float×10: 理论竞争, 实际无害, 不改

TODO审查→无需修改:
- TODO-1: centroid数据已验证, 7/8场景ratio≤1.6x, LMS自修正
- TODO-2: NLMS辨识完整冲激响应, FIR系数自动编码极性, 不存在符号问题
- TODO-3: 离线/在线信号链天生不同, LMS功率归一化自动补偿, 非bug

文档:
- 量产阻塞项总览 (8项真正阻塞量产的问题)
- P2-4严重度🟡→🔴 (硬件看门狗, 量产必须)
- TODO-5/CRC-2严重度提升
- 多名条目同步: ✅/—/已修复标记

+++ include/fxnlms_mimo.h
@@ -11,7 +11,7 @@ typedef struct {
     int          E, S, L;
     float        step_size;
     float        leak;
-    volatile int freeze_lms; /* 发散检测: 1=冻结梯度更新, 0=正常 */
+    volatile long freeze_lms; /* 发散检测: 1=冻结梯度更新, 0=正常 (long=LONG, Interlocked兼容) */
 } fxnlms_mimo_t;
 
 int  fxnlms_init(fxnlms_mimo_t *fx, int E, int S, int L,
